+++ test/Task.ts
@@ -160,7 +160,7 @@ describe('Task', () => {
             return n
           })
         )
-      const as = RA.makeBy(4, append)
+      const as = pipe(4, RA.makeBy(append))
       U.deepStrictEqual(await pipe(as, _.traverseReadonlyArrayWithIndex(SK))(), [0, 1, 2, 3])
       U.deepStrictEqual(log, [0, 2, 1, 3])
     })
@@ -174,7 +174,7 @@ describe('Task', () => {
             return n
           })
         )
-      const as = RA.makeBy(4, append)
+      const as = pipe(4, RA.makeBy(append))
       U.deepStrictEqual(await pipe(as, _.traverseReadonlyArrayWithIndexSeq(SK))(), [0, 1, 2, 3])
       U.deepStrictEqual(log, [0, 1, 2, 3])
     })